Move loadOME* methods to ome.ts. Fix click on Well
